Makefile: compatible beyond Broadcom router build
Now the locations of dependent libraries could be customized
with the global variables, which indicate the new places.

With variable ROOTDIR, the locations of dependent libraries could
be customized and compatible with the original implementation.

The change also supports to build the source code beyond the source
code directory and refer to outside source and binrary files.

Change-Id: Icf0bbcc2ab596fc29b3bc1e36fd028f6c397c34d